pyproject.toml
[project]
name = "tetris"
version = "1.0.0"
description = ""
authors = [
    {name = "Alexandr Mikhalchenkov",email = "mikhalchenkov22@gmail.com"}
]
license = {text = "MIT License (https://opensource.org/licenses/MIT)"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "python-keycloak (>=5.3.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
